requirements.txt
Flask==2.3.3
mysql-connector-python==8.0.30
pandas==2.0.3
openpyxl==3.1.0
face_recognition==1.3.0
Flask-Cors==3.0.10
